feat: :lipstick: pastas organizadas
separação das pastas para preparar para implementar os eventos de domínio
