+++ package.json
@@ -14,9 +14,11 @@
     "tsc": "tsc"
   },
   "dependencies": {
+    "@types/uuid": "^10.0.0",
     "reflect-metadata": "^0.2.2",
     "sequelize": "^6.37.3",
     "sequelize-typescript": "^2.1.6",
-    "sqlite3": "^5.1.7"
+    "sqlite3": "^5.1.7",
+    "uuid": "^10.0.0"
   }
 }
