+++ src/domain/@shared/event/event-dispatcher.interface.ts
@@ -1,0 +1,9 @@
+import EventHandlerInterface from "./event-handler.interface";
+import EventInterface from "./event.interface";
+
+export default interface EventDispatcherInterface {
+    notify(event: EventInterface): void
+    register(eventName: string, eventHandler: EventHandlerInterface): void
+    unregister(eventName: string, eventHandler: EventHandlerInterface): void
+    unregisterAll(): void
+}
