+++ src/domain/@shared/event/event-handler.interface.ts
@@ -1,0 +1,5 @@
+import EventInterface from "./event.interface";
+
+export default interface EventHandlerInterface<T extends EventInterface=EventInterface> {
+    handle(event: T): void
+}
