+++ src/domain/@shared/event/event.interface.ts
@@ -1,0 +1,4 @@
+export default interface EventInterface {
+    dateTimeOcurred: Date
+    eventData: any
+}
