+++ src/domain/checkout/repository/order-repository.interface.ts
@@ -1,0 +1,5 @@
+import Order from "../entity/order";
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+
+export default interface OrderRepositoryInterface extends RepositoryInterface<Order> {
+}
