+++ src/domain/customer/factory/customer.factory.spec.ts
@@ -1,0 +1,22 @@
+import CustomerFactory from "./customer.factory";
+import Address from "../value-object/address";
+
+describe("Customer factory unit test", () => {
+  it("should create a customer", () => {
+    let customer = CustomerFactory.create("John");
+
+    expect(customer.id).toBeDefined();
+    expect(customer.name).toBe("John");
+    expect(customer.address).toBeUndefined();
+  });
+
+  it("should create a customer with an address", () => {
+    const address = new Address("Street", 1, "13330-250", "São Paulo");
+
+    let customer = CustomerFactory.createWithAddress("John", address);
+
+    expect(customer.id).toBeDefined();
+    expect(customer.name).toBe("John");
+    expect(customer.address).toBe(address);
+  });
+});
