+++ src/domain/customer/repository/customer-repository.interface.ts
@@ -1,0 +1,5 @@
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+import Customer from "../entity/customer";
+
+export default interface CustomerRepositoryInterface extends RepositoryInterface<Customer> {
+}
