+++ src/domain/product/entity/product.interface.ts
@@ -1,0 +1,5 @@
+export default interface ProductInterface {
+    get id(): string
+    get name(): string
+    get price(): number
+}
