+++ src/domain/product/event/product-created.event.ts
@@ -1,0 +1,11 @@
+import EventInterface from "../../@shared/event/event.interface";
+
+export default class ProductCreatedEvent implements EventInterface {
+    dateTimeOcurred: Date;
+    eventData: any;
+    
+    constructor(eventData: any) {
+        this.dateTimeOcurred = new Date()
+        this.eventData = eventData
+    }
+}
