+++ src/domain/product/repository/product-repository.interface.ts
@@ -1,0 +1,5 @@
+import Product from "../entity/product";
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+
+export default interface ProductRepositoryInterface extends RepositoryInterface<Product> {
+}
