+++ src/domain/product/service/product.service.ts
@@ -1,0 +1,10 @@
+import Product from "../entity/product";
+
+export default class ProductService {
+    static increasePrice(products: Product[], percentage: number): Product[] {
+      products.forEach((product) => {
+        product.changePrice((product.price * percentage) / 100 + product.price);
+      });
+      return products;
+    }
+}
