--- src/domain/repository/customer-repository.interface.ts
@@ -1,5 +1,0 @@
-import Customer from "../entity/customer";
-import RepositoryInterface from "./repository-interface";
-
-export default interface CustomerRepositoryInterface extends RepositoryInterface<Customer> {
-}
