--- src/domain/repository/order-repository.interface.ts
@@ -1,5 +1,0 @@
-import Order from "../entity/order";
-import RepositoryInterface from "./repository-interface";
-
-export default interface OrderRepositoryInterface extends RepositoryInterface<Order> {
-}
