--- src/domain/repository/product-repository.interface.ts
@@ -1,5 +1,0 @@
-import Product from "../entity/product";
-import RepositoryInterface from "./repository-interface";
-
-export default interface ProductRepositoryInterface extends RepositoryInterface<Product> {
-}
