+++ src/infrastructure/order/repository/sequelize/order-item.model.ts
@@ -1,0 +1,47 @@
+import { BelongsTo, Column, ForeignKey, Model, PrimaryKey, Table } from "sequelize-typescript";
+import ProductModel from "../../../product/repository/sequelize/product.model";
+import OrderModel from "./order.model";
+import { NonAttribute } from "sequelize";
+
+@Table({
+    tableName: "order_items",
+    timestamps: false,
+})
+export default class OrderItemModel extends Model {
+    @PrimaryKey
+    @Column
+    declare id: string
+
+    @ForeignKey(() => ProductModel)
+    @Column({ allowNull: false})
+    declare product_id: string
+
+    @BelongsTo(() => ProductModel)
+    declare product: ProductModel
+
+    @ForeignKey(() => OrderModel)
+    @Column({ allowNull: false})
+    declare order_id: string
+
+    @BelongsTo(() => OrderModel,
+    {
+        constraints: false,
+        foreignKey: "order_id",
+        as: "items",
+        keyType: "hasMany"
+        // inverse: {
+        //     as: "items",
+        //     type: "hasMany"
+        // },
+    })
+    declare order?: NonAttribute<OrderModel>
+
+    @Column({ allowNull: false})
+    declare quantity: number
+
+    @Column({ allowNull: false})
+    declare name: string
+
+    @Column({ allowNull: false})
+    declare price: number
+}
