+++ src/infrastructure/product/repository/sequelize/product.repository.ts
@@ -1,0 +1,47 @@
+import Product from "../../../../domain/product/entity/product";
+import ProductRepositoryInterface from "../../../../domain/product/repository/product-repository.interface";
+import ProductModel from "./product.model";
+
+export default class ProductRepository implements ProductRepositoryInterface {
+    async create(entity: Product): Promise<void> {
+        try {
+            await ProductModel.create({
+                id: entity.id,
+                name: entity.name,
+                price: entity.price,
+            })
+        } catch (error) {
+            console.log(error)
+            throw error
+        }
+    }
+    async update(entity: Product): Promise<void> {
+        await ProductModel.update({
+            name: entity.name,
+            price: entity.price,
+        },{
+            where: {
+                id: entity.id
+            }
+        })
+    }
+    async find(id: string): Promise<Product> {
+        const productModel = await ProductModel.findOne({
+            where: {
+                id: id,
+            }
+        })
+        return new Product(
+            productModel.id,
+            productModel.name,
+            productModel.price,
+        )
+    }
+    async findAll(): Promise<Product[]> {
+        const productModels = await ProductModel.findAll()
+        return productModels.map(productModel => 
+            new Product(productModel.id, productModel.name, productModel.price)
+        )
+    }
+
+}
